% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stepfun.R
\name{busy_server_stepfun}
\alias{busy_server_stepfun}
\title{busy server stepfun (and server utilisation)}
\usage{
busy_server_stepfun(systemlength_df, servers)
}
\arguments{
\item{systemlength_df}{systemlength_df from queue_step}

\item{servers}{server stepfunction}
}
\description{
busy server stepfun (and server utilisation)
}
\examples{
n <- 100
arrivals <- cumsum(rexp(n, 2.5))
service <- rexp(n)
x <- c(10, 40)
y <- c(1, 2, 1)
queue_obj <- queue_step(arrivals, service, servers = as.server.stepfun(x,y))

output <- busy_server_stepfun(queue_obj$systemlength_df, stepfun(x,y))

plot(output$busy_stepfun)
plot(output$util_stepfun)
}
