% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_batches}
\alias{create_batches}
\title{Creates batches of customer arrivals from a dataframe within a \code{dplyr::do} command}
\usage{
create_batches(data, arrival_dist, service_rate = NULL, time = 0)
}
\arguments{
\item{data}{a named list object.}

\item{arrival_dist}{a distribution whose support is strictly positive. Either as an object or a non-empty character string. It represents the distribution of arrival times.}

\item{service_rate}{a strictly positive number representing the rate parameter in the exponential distribution for the service times.}

\item{time}{a number greater than or equal to zero.}
}
\description{
Creates batches of customer arrivals from a dataframe within a \code{dplyr::do} command
}

