% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{depart}
\alias{depart}
\title{get departure times from \code{queue_list} object}
\usage{
depart(x)
}
\arguments{
\item{x}{an \code{queue_list} object}
}
\value{
departure times
}
\description{
get departure times from \code{queue_list} object
}
\examples{
arrivals <- cumsum(rexp(10))
service <- rexp(10)
queue_obj <- queue_step(arrivals, service)

depart(queue_obj)
queue_obj$departures_df$departures
}
