% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head.queue_list}
\alias{head.queue_list}
\title{head method for objects of class \code{queue_list}}
\usage{
\method{head}{queue_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{queue_list} produced by the \code{\link{queue_step}} function.}

\item{...}{further arguments to be passed to other methods}
}
\description{
head method for objects of class \code{queue_list}
}

