% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrate_stepfun_interval}
\alias{integrate_stepfun_interval}
\title{Integrate step function over interval}
\usage{
integrate_stepfun_interval(x, y, from = 0, to = 1000)
}
\arguments{
\item{x}{numeric vector giving the times of changes in number of servers.}

\item{y}{numeric vector one longer than \code{x} giving the number of servers
available between x values.}

\item{from}{start of integration}

\item{to}{end of intergration}
}
\description{
Integrate step function over interval
}
\examples{
x_knots <- c(7.5, 15, 25, 40)
y <- c(5, 4, 2, 6, 4)

curve(stepfun(x_knots, y)(x), from = 0, to = 100)
abline(v = 10, col = "red")
abline(v = 30, col = "red")

integrate_stepfun_interval(x_knots, y, from = 10, to = 30)

# Time-average number of servers available in interval
integrate_stepfun_interval(x_knots, y, from = 10, to = 30) / (30 - 10)
}
