% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{lag_step}
\alias{lag_step}
\title{Add lag to vector of arrival times.}
\usage{
lag_step(arrival_df, service)
}
\arguments{
\item{arrival_df}{A dataframe with column names ID and times . The ID column is a key
for the customers. The times column is of class \code{numeric} and represents the
arrival times of the customers.}

\item{service}{A vector of service times with the same ordering as arrival_df.}
}
\value{
A vector of response times for the input of arrival times and service times.
}
\description{
Add lag to vector of arrival times.
}
\seealso{
\code{\link[queuecomputer]{wait_step}}, \code{\link[queuecomputer]{queue_step}}
}

