% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_max}
\alias{n_max}
\title{Return \code{n}th maximum of vector}
\usage{
n_max(input, n, tail_variable = 500)
}
\arguments{
\item{input}{A vector of numbers.}

\item{n}{An integer.}
}
\value{
\code{n} th maximum of \code{input}
}
\description{
Return \code{n}th maximum of vector
}
\examples{
n_max(c(3,1,7),2)
}
\seealso{
wait_step
}

