% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_function}
\alias{next_function}
\title{Return next available time for server given current departure times.}
\usage{
next_function(sf, time)
}
\arguments{
\item{sf}{A step function with heights either 0 or 1.}

\item{time}{the time in question.}
}
\value{
Next available time for server in question.
}
\description{
Return next available time for server given current departure times.
}

