% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_stepfun.R
\name{performance_output}
\alias{performance_output}
\title{Performance data}
\usage{
performance_output(arrivals, service = 0, departures, servers,
  epsilon = 1e-10)
}
\arguments{
\item{arrivals}{arrivals}

\item{service}{service}

\item{departures}{departures}

\item{servers}{server object}

\item{epsilon}{epsilon}
}
\description{
Performance data
}
\examples{
n <- 300
arrivals <- cumsum(rexp(n, 2.8))
service <- rexp(n)
x <- c(10, 40, 100)
y <- c(1, 5, 2, 5)
queue_obj <- queue_step(arrivals, service, servers = as.server.stepfun(x,y))

output <- performance_output(arrivals, service, queue_obj$departures, queue_obj$servers_input)

plot(output$busy_df, type = "l")
plot(output$util_df, type = "l")
}
