% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.queue_list}
\alias{plot.queue_list}
\title{plot method}
\usage{
\method{plot}{queue_list}(x, server_same = TRUE)
}
\arguments{
\item{x}{an object of class \code{queue_list}}
}
\description{
plot method
}
\examples{
n_customers <- 1e3
arrival_rate <- 0.8
service_rate <- 1
arrivals <- cumsum(rexp(n_customers, arrival_rate))
service <- rexp(n_customers, service_rate)
queue_obj <- QDC(arrivals, service, servers = 1)
plot(queue_obj)
}

