% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.queue_list}
\alias{print.queue_list}
\title{print method for objects of class \code{queue_list}}
\usage{
\method{print}{queue_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{queue_list} produced by the \code{\link{queue_step}} function.}

\item{...}{further arguments to be passed to other methods}
}
\description{
print method for objects of class \code{queue_list}
}
