% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{print.summary_queue_df}
\alias{print.summary_queue_df}
\title{Print method for output of \code{summary.queue_df}.}
\usage{
\method{print}{summary_queue_df}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary_queue_df}, the result of a call to \code{summary.queue_df()}.}

\item{...}{futher arguments to be passed to or from other methods.}
}
\value{
A list of performance statistics for the queue. "Mean waiting time": The mean time each customer had to wait in queue for service. "Mean response time": The mean time that each customer spends in the system (departure time - arrival time). "Utilization factor": The ratio of available time for all servers and time all servers were used. It can be greater than one if a customer arrives near the end of a shift and keeps a server busy. "Queue Lengths" and "System Lengths", the proportion of time for each queue length or number of customers in the system.
}
\description{
Print method for output of \code{summary.queue_df}.
}

