% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{QDC}
\alias{QDC}
\title{Compute the queue departure times of customers given a set of arrival times, a set of service times, and a resource schedule.}
\usage{
QDC(arrivals, service, servers = 1, labels = NULL, adjust = 1)
}
\arguments{
\item{arrivals}{numeric vector of non-negative arrival times}

\item{service}{numeric vector of service times with the same ordering as arrival_df.}

\item{servers}{a non-zero natural number, an object of class \code{server.stepfun}
or an object of class \code{server.list}.}

\item{labels}{character vector of customer labels.}

\item{adjust}{non-negative number, an adjustment parameter for scaling the service times.}
}
\value{
A vector of response times for the input of arrival times and service times.
}
\description{
Compute the queue departure times of customers given a set of arrival times, a set of service times, and a resource schedule.
}
\examples{
n_customers <- 1e3
arrival_rate <- 0.8
service_rate <- 1

arrivals <- cumsum(rexp(n_customers, arrival_rate))
service <- rexp(n_customers, service_rate)
queue_obj <- QDC(arrivals, service, servers = 1)
queue_obj
summary(queue_obj)
}

