% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{queue}
\alias{queue}
\title{Compute the departure times of a set of customers in a queue from their arrival and service times.}
\usage{
queue(arrivals, service, servers = 1, serveroutput = FALSE)
}
\arguments{
\item{arrivals}{numeric vector of non-negative arrival times}

\item{service}{numeric vector of non-negative service times}

\item{servers}{a non-zero natural number, an object of class \code{server.stepfun}
or an object of class \code{server.list}.}

\item{serveroutput}{boolean whether the server used by each customer should be returned.}
}
\description{
Compute the departure times of a set of customers in a queue from their arrival and service times.
}

