% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{queue_step}
\alias{queue_step}
\title{Compute the vector of response times from a queue with a vector of arrival times and service times.}
\usage{
queue_step(arrival_df, Number_of_queues = stepfun(1, c(1, 1)), service)
}
\arguments{
\item{arrival_df}{A dataframe with column names ID and times . The ID column is a key
for the customers. The times column is of class \code{numeric} and represents the
arrival times of the customers.}

\item{Number_of_queues}{A step function which describes the number of service desks open}

\item{service}{A vector of service times with the same ordering as arrival_df}
}
\value{
A vector of response times for the input of arrival times and service times
}
\description{
Compute the vector of response times from a queue with a vector of arrival times and service times.
}
\examples{
arrival_df <- data.frame(ID = c(1:100), times = rlnorm(100, meanlog = 3))
service <- rlnorm(100)
queue_step(arrival_df = arrival_df, service = service)
queue_step(arrival_df = arrival_df,
    Number_of_queues = stepfun(c(15,30,50), c(1,3,1,10)), service = service)

ord <- order(arrival_df$times)
cbind(arrival_df[ord,], service[ord],
    queue_step(arrival_df = arrival_df,
    Number_of_queues = stepfun(c(15,30,50), c(1,3,1,10)), service = service)$times[ord])
}
\references{
Sutton, C., & Jordan, M. I. (2011). Bayesian inference for queueing networks and modeling of internet services. The Annals of Applied Statistics, 254-282, page 258.
}
\seealso{
wait_step, lag_step
}

