% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_bags}
\alias{reduce_bags}
\title{Return arrival vector of bags dataset.}
\usage{
reduce_bags(bagdataset, number_of_passengers)
}
\arguments{
\item{bagdataset}{A dataset of bags.}

\item{number_of_passengers}{The total number of passengers all the bags could belong to.}
}
\value{
Maximum times for each passenger ID
}
\description{
Return arrival vector of bags dataset.
}
\examples{
bags <- rpois(100,1)
bags.df <- data.frame(BagID = 1:sum(bags),
    ID = rep(1:100, bags), times = rlnorm(sum(bags), meanlog = 2))
arrivals2 <- reduce_bags(bags.df, 100)
}
\seealso{
wait_step
}

