% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{server_list}
\alias{server_list}
\title{Creates a \code{"server_list"} object from a list of times and starting availability.}
\usage{
server_list(times, init)
}
\arguments{
\item{times}{list of vectors giving change times for each server.}

\item{init}{vector of 1s and 0s which represents whether the server starts availabile \code{(1)} or unavailable \code{(0)}.}
}
\value{
\code{"server_list"} object.
}
\description{
Creates a \code{"server_list"} object from a list of times and starting availability.
}
\examples{
# Create a server_list with the first server available anytime before time 10,
# and the second server available between time 15 and time 30.
server_list(list(10, c(15,30)), c(1,0))
}
\seealso{
\code{\link{server_split}}, \code{\link{queue_step}}
}

