% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{summary.queue_df}
\alias{summary.queue_df}
\title{Compute resource statistics from queue_df object}
\usage{
\method{summary}{queue_df}(queue_df)
}
\description{
Compute resource statistics from queue_df object
}

