% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{summary.queue_list}
\alias{summary.queue_list}
\title{Summary method for queue_list object}
\usage{
\method{summary}{queue_list}(object, ...)
}
\arguments{
\item{object}{an object of class \code{queue_list}, the result of a call to \code{queue_step}.}

\item{...}{futher arguments to be passed to or from other methods.}
}
\description{
Summary method for queue_list object
}

