% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_nmax}
\alias{which_nmax}
\title{Find nth maximum of vector.}
\usage{
which_nmax(input, n)
}
\arguments{
\item{input}{A vector.}

\item{n}{An integer.}
}
\value{
The position of the \code{n} th maximum of \code{input}
}
\description{
Find nth maximum of vector.
}
\examples{
which_nmax(c(40,50,30),1)
which_nmax(c(40,50,30),2)
which_nmax(c(40,50,30),3)
}

