% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{which_nmin}
\alias{which_nmin}
\title{Find nth minimum of vector.}
\usage{
which_nmin(input, n)
}
\arguments{
\item{input}{A vector.}

\item{n}{An integer.}
}
\value{
The position of the \code{n} th minimum of \code{input}
}
\description{
Find nth minimum of vector.
}
\examples{
which_nmin(c(40,50,30),1)
which_nmin(c(40,50,30),2)
which_nmin(c(40,50,30),3)
}
\seealso{
\code{\link{which_nmax}}
}

