% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{queue_fast}
\alias{queue_fast}
\title{Compute the vector of response times from a queue with a vector of arrival times and service times.}
\usage{
queue_fast(arrival_df, Number_of_servers = 1, service)
}
\arguments{
\item{arrival_df}{A dataframe with column names ID and times . The ID column is a key
for the customers. The times column is of class \code{numeric} and represents the
arrival times of the customers.}

\item{Number_of_servers}{The number of servers in the queue model.}

\item{service}{A vector of service times with the same ordering as arrival_df.}
}
\value{
A vector of response times for the input of arrival times and service times
}
\description{
Compute the vector of response times from a queue with a vector of arrival times and service times.
}
\examples{
set.seed(700)
arrival_df <- data.frame(ID = c(1:100), times = rlnorm(100, meanlog = 3))
service <- rlnorm(100)


firstqueue <- queue_fast(arrival_df = arrival_df, service = service)
secondqueue <- queue_fast(arrival_df = arrival_df,
    Number_of_servers = 2, service = service)

curve(ecdf(arrival_df$times)(x) * 100 , from = 0, to = 200,
    xlab = "time", ylab = "Number of customers")
curve(ecdf(firstqueue$times)(x) * 100 , add = TRUE, col = "red")
curve(ecdf(secondqueue$times)(x) * 100, add = TRUE, col = "blue")
legend(100,40, legend = c("Customer input - arrivals",
    "Customer output - firstqueue",
    "Customer output - secondqueue"),
    col = c("black","red","blue"), lwd = 1, cex = 0.8
)

# Queue lengths
ecdf(arrival_df$times)(c(1:200))*100 - ecdf(firstqueue$times)(c(1:200))*100
ecdf(arrival_df$times)(c(1:200))*100 - ecdf(secondqueue$times)(c(1:200))*100

ord <- order(arrival_df$times)
cbind(arrival_df[ord,], service[ord],
    secondqueue$times[ord])
}
\seealso{
wait_step, lag_step
}

