% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_queuefunctions.R
\name{queue_step}
\alias{queue_step}
\title{Compute the departure times of customers given: a set of arrival and service times, and a resource schedule.}
\usage{
queue_step(arrival_df, service, servers = 1)
}
\arguments{
\item{arrival_df}{dataframe with column names ID and times . The ID column is a key
for the customers. The times column is of class \code{numeric} and represents the
arrival times of the customers.}

\item{service}{vector of service times with the same ordering as arrival_df.}

\item{servers}{an integer or an object of class \code{"server.list"}.}
}
\value{
A vector of response times for the input of arrival times and service times.
}
\description{
Compute the departure times of customers given: a set of arrival and service times, and a resource schedule.
}
\examples{
set.seed(1L)
n_customers <- 100
arrival_df <- data.frame(ID = c(1:n_customers), times = rlnorm(n_customers, meanlog = 3))
service <- rlnorm(n_customers)

server_list <- server_split(c(15),c(1,2))

firstqueue <- queue_step(arrival_df = arrival_df, service = service)
secondqueue <- queue_step(arrival_df = arrival_df,
    servers = server_list, service = service)

curve(ecdf(arrival_df$times)(x) * n_customers , from = 0, to = 200,
    xlab = "time", ylab = "Number of customers")
curve(ecdf(firstqueue$times)(x) * n_customers , add = TRUE, col = "red")
curve(ecdf(secondqueue$times)(x) * n_customers, add = TRUE, col = "blue")
legend(100,40, legend = c("Customer input - arrivals",
    "Customer output - firstqueue",
    "Customer output - secondqueue"),
    col = c("black","red","blue"), lwd = 1, cex = 0.8
)

# Queue lengths
ecdf(arrival_df$times)(c(1:200))*100 - ecdf(firstqueue$times)(c(1:200))*100
ecdf(arrival_df$times)(c(1:200))*100 - ecdf(secondqueue$times)(c(1:200))*100

ord <- order(arrival_df$times)
cbind(arrival_df[ord,], service[ord],
    secondqueue$times[ord], attr(secondqueue, "queue")[ord])
}
\seealso{
\code{\link[queuecomputer]{wait_step}}, \code{\link[queuecomputer]{lag_step}}, \code{\link[queuecomputer]{as.server.list}}, \code{\link[queuecomputer]{server_split}}
}

