% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{server_split}
\alias{server_split}
\title{Create a \code{quick.q} object with a roster of times and number of available servers.}
\usage{
server_split(x, y)
}
\arguments{
\item{x}{numeric vector giving the times of changes in number of servers.}

\item{y}{numeric vector giving the number of servers available between x values.}
}
\value{
A \code{list} and \code{quick.q} object with x and y as elements.
}
\description{
Create a \code{quick.q} object with a roster of times and number of available servers.
}
\examples{

servers <- server_split(c(15,30,50), c(0, 1, 3, 2))
servers

}
\seealso{
\code{\link{as.server.list}}, \code{\link{queue_step}}
}

