% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{server_split}
\alias{server_split}
\title{Create a \code{"server_list"} object with a roster of times and number of available servers.}
\usage{
server_split(x, y, quick = TRUE)
}
\arguments{
\item{x}{numeric vector giving the times of changes in number of servers.}

\item{y}{numeric vector giving the number of servers available between x values.}

\item{quick}{logical should the quick method be used?}
}
\value{
List of height 1 step functions for input into queue_step.
}
\description{
Create a \code{"server_list"} object with a roster of times and number of available servers.
}
\examples{

server_split(c(15,30,50), c(0, 1, 3, 2))
}
\seealso{
\code{\link{as.server.list}}, \code{\link{queue_step}}
}

