% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{server_split}
\alias{server_split}
\title{Makes it easier to input number of servers into queue_step.}
\usage{
server_split(x, y)
}
\arguments{
\item{x}{numeric vector giving the knots or jump locations of the step function for stepfun().}

\item{y}{numeric vector giving the knots or jump locations of the step function for stepfun(). Must be integers.}
}
\value{
List of height 1 step functions for input into queue_step.
}
\description{
Makes it easier to input number of servers into queue_step.
}
\seealso{
queue_step()
}

